Conditionally configure Swagger based on environment
The code now conditionally configures Swagger and Swagger UI based on the environment. In the development environment, `app.UseDeveloperExceptionPage()` is used along with Swagger and Swagger UI. In non-development environments, only Swagger and Swagger UI are configured. The redundant unconditional calls to `app.UseSwagger()` and `app.UseSwaggerUI()` have been removed.