Add PUT endpoint for updating SubTipo and new DTO class
A new HTTP PUT endpoint has been added to the SubTiposController class in the SiAB.API.Controllers.Misc namespace. This endpoint allows for updating an existing SubTipo entity by its ID. The method Update takes an id from the route and an UpdateSubTipoDto object from the request body. It retrieves the entity by ID, updates its properties, and saves the changes to the repository. If the entity is not found, it returns a NotFound response.

Additionally, a new DTO class UpdateSubTipoDto has been added in the SiAB.Core.DTO.Misc namespace. This class inherits from UpdateNamedEntityDto and includes an additional property TipoId of type int.