Add new endpoint and fix property mappings
A new endpoint `adjuntar-formulario-53` has been added to the `TransaccionesController` class in the `SiAB.API.Controllers.Belico` namespace. This endpoint is decorated with the `CreateAuditableFilter` service filter and is designed to handle HTTP POST requests. It accepts a `AdjuntarFormularioTransaccionDto` object as input and calls the `SaveFormulario53` method of the `TransaccionRepository` to save the form data.

In the `TrasaccionRepository` class within the `SiAB.Infrastructure.Repositories.Belico` namespace, the properties `TipoOrigen`, `Origen`, `TipoDestino`, and `Destino` of the `transaccion` object have been swapped. Specifically, `TipoOrigen` now maps to `TipoCargoCredito`, `Origen` to `Credito`, `TipoDestino` to `TipoCargoDebito`, and `Destino` to `Debito`. This change corrects the mapping of transaction types and their corresponding values.